@@ -1,31 +1,48 @@
 {
-	"name": "examples",
-	"version": "0.0.1",
-	"private": true,
-	"scripts": {
-		"start": "node node_modules/react-native/local-cli/cli.js start",
-		"watch": "node ./scripts/watch-and-copy-src.js",
-		"packager": "sh ./node_modules/react-native/packager/packager.sh --root ./node_modules/react-native-swiper",
-		"dev": "concurrently 'npm run watch' 'npm run packager'",
-		"test": "jest"
-	},
-	"dependencies": {
-		"react": "16.0.0-alpha.12",
-		"react-native": "0.48.4",
-		"react-native-swiper": "../"
-	},
-	"devDependencies": {
-		"babel-jest": "20.0.3",
-		"babel-preset-react-native": "2.1.0",
-		"concurrently": "^2.2.0",
-		"fs-extra": "^0.30.0",
-		"jest": "20.0.4",
-		"minimatch": "^3.0.2",
-		"node-watch": "^0.4.0",
-		"react-test-renderer": "16.0.0-alpha.12",
-		"rimraf": "^2.5.4"
-	},
-	"jest": {
-		"preset": "react-native"
-	}
+  "name": "examples",
+  "version": "0.0.1",
+  "private": true,
+  "scripts": {
+    "start": "rm -rf node_modules/react-native-swiper/examples && react-native start || react-native start",
+    "test": "jest",
+    "lint": "eslint ."
+  },
+  "dependencies": {
+    "react": "16.8.6",
+    "react-model": "^3.0.0",
+    "react-native": "0.60.5",
+    "react-native-gesture-handler": "^1.4.1",
+    "react-native-reanimated": "^1.3.0",
+    "react-native-screens": "^1.0.0-alpha.23",
+    "react-native-swiper": "../",
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.9.3"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.6.0",
+    "@babel/runtime": "^7.6.0",
+    "@react-native-community/eslint-config": "^0.0.5",
+    "@types/react": "^16.9.2",
+    "@types/react-native": "^0.60.15",
+    "babel-jest": "^24.9.0",
+    "detox": "^14.3.4",
+    "eslint": "^6.4.0",
+    "jest": "^24.9.0",
+    "metro-react-native-babel-preset": "^0.56.0",
+    "react-test-renderer": "16.8.6"
+  },
+  "jest": {
+    "preset": "react-native"
+  },
+  "detox": {
+    "test-runner": "jest",
+    "configurations": {
+      "ios.sim.debug": {
+        "binaryPath": "ios/build/examples/Build/Products/Debug-iphonesimulator/examples.app",
+        "build": "xcodebuild -workspace ios/examples.xcworkspace -scheme examples -configuration Debug -sdk iphonesimulator -derivedDataPath ios/build",
+        "type": "ios.simulator",
+        "name": "iPhone 7"
+      }
+    }
+  }
 }
