@@ -216,7 +216,7 @@ export default class extends Component {
     this.loopJumpTimer && clearTimeout(this.loopJumpTimer)
   }
 
-  componentWillUpdate(nextProps, nextState) {
+  UNSAFE_componentWillUpdate(nextProps, nextState) {
     // If the index has changed, we notify the parent via the onIndexChanged callback
     if (this.state.index !== nextState.index)
       this.props.onIndexChanged(nextState.index)
@@ -228,9 +228,13 @@ export default class extends Component {
       this.autoplay()
     }
     if (this.props.children !== prevProps.children) {
-      this.setState(
-        this.initState({ ...this.props, index: this.state.index }, true)
-      )
+      if (this.props.loadMinimal && Platform.OS === 'ios') {
+        this.setState({ ...this.props, index: this.state.index })
+      } else {
+        this.setState(
+          this.initState({ ...this.props, index: this.state.index }, true)
+        )
+      }
     }
   }
 
@@ -282,7 +286,7 @@ export default class extends Component {
     }
 
     initState.offset[initState.dir] =
-      initState.dir === 'y' ? height * props.index : width * props.index
+      initState.dir === 'y' ? initState.height * props.index : initState.width * props.index
 
     this.internals = {
       ...this.internals,
@@ -312,19 +316,18 @@ export default class extends Component {
 
     // only update the offset in state if needed, updating offset while swiping
     // causes some bad jumping / stuttering
-    if (
-      !this.state.offset ||
-      width !== this.state.width ||
-      height !== this.state.height
-    ) {
+    if (!this.state.offset) {
       state.offset = offset
     }
 
     // related to https://github.com/leecade/react-native-swiper/issues/570
     // contentOffset is not working in react 0.48.x so we need to use scrollTo
     // to emulate offset.
-    if (this.initialRender && this.state.total > 1) {
+    if(this.state.total > 1) {
       this.scrollView.scrollTo({ ...offset, animated: false })
+    }
+	
+    if (this.initialRender) {
       this.initialRender = false
     }
 
@@ -336,10 +339,34 @@ export default class extends Component {
     const i = this.state.index + (this.props.loop ? 1 : 0)
     const scrollView = this.scrollView
     this.loopJumpTimer = setTimeout(
-      () =>
-        scrollView.setPageWithoutAnimation &&
-        scrollView.setPageWithoutAnimation(i),
-      50
+      () => {
+        if (scrollView.setPageWithoutAnimation) {
+          scrollView.setPageWithoutAnimation(i)
+        } else {
+          if (this.state.index === 0) {
+            scrollView.scrollTo(
+              this.props.horizontal === false
+                ? { x: 0, y: this.state.height, animated: false }
+                : { x: this.state.width, y: 0, animated: false }
+            )
+          } else if (this.state.index === this.state.total - 1) {
+            this.props.horizontal === false
+              ? this.scrollView.scrollTo({
+                  x: 0,
+                  y: this.state.height * this.state.total,
+                  animated: false
+                })
+              : this.scrollView.scrollTo({
+                  x: this.state.width * this.state.total,
+                  y: 0,
+                  animated: false
+                })
+          }
+        }
+      },
+      // Important Parameter
+      // ViewPager 50ms, ScrollView 300ms
+      scrollView.setPageWithoutAnimation ? 50 : 300
     )
   }
 
@@ -404,11 +431,10 @@ export default class extends Component {
     this.updateIndex(e.nativeEvent.contentOffset, this.state.dir, () => {
       this.autoplay()
       this.loopJump()
-
-      // if `onMomentumScrollEnd` registered will be called here
-      this.props.onMomentumScrollEnd &&
-        this.props.onMomentumScrollEnd(e, this.fullState(), this)
     })
+    // if `onMomentumScrollEnd` registered will be called here
+    this.props.onMomentumScrollEnd &&
+      this.props.onMomentumScrollEnd(e, this.fullState(), this)
   }
 
   /*
@@ -439,41 +465,11 @@ export default class extends Component {
   updateIndex = (offset, dir, cb) => {
     const state = this.state
     // Android ScrollView will not scrollTo certain offset when props change
-    const callback = async () => {
-      cb()
-      if (Platform.OS === 'android') {
-        if (this.state.index === 0) {
-          this.props.horizontal
-            ? this.scrollView.scrollTo({
-                x: state.width,
-                y: 0,
-                animated: false
-              })
-            : this.scrollView.scrollTo({
-                x: 0,
-                y: state.height,
-                animated: false
-              })
-        } else if (this.state.index === this.state.total - 1) {
-          this.props.horizontal
-            ? this.scrollView.scrollTo({
-                x: state.width * this.state.total,
-                y: 0,
-                animated: false
-              })
-            : this.scrollView.scrollTo({
-                x: 0,
-                y: state.height * this.state.total,
-                animated: false
-              })
-        }
-      }
-    }
     let index = state.index
     if (!this.internals.offset)
       // Android not setting this onLayout first? https://github.com/leecade/react-native-swiper/issues/582
       this.internals.offset = {}
-    const diff = offset[dir] - this.internals.offset[dir]
+    const diff = offset[dir] - (this.internals.offset[dir] || 0)
     const step = dir === 'x' ? state.width : state.height
     let loopJump = false
 
@@ -514,14 +510,14 @@ export default class extends Component {
         newState.offset = { x: 0, y: 0 }
         newState.offset[dir] = offset[dir] + 1
         this.setState(newState, () => {
-          this.setState({ offset: offset }, callback)
+          this.setState({ offset: offset }, cb)
         })
       } else {
         newState.offset = offset
-        this.setState(newState, callback)
+        this.setState(newState, cb)
       }
     } else {
-      this.setState(newState, callback)
+      this.setState(newState, cb)
     }
   }
 
@@ -853,8 +849,12 @@ export default class extends Component {
       pages = pages.map((page, i) => {
         if (loadMinimal) {
           if (
-            i >= index + loopVal - loadMinimalSize &&
-            i <= index + loopVal + loadMinimalSize
+            (i >= index + loopVal - loadMinimalSize &&
+              i <= index + loopVal + loadMinimalSize) ||
+            // The real first swiper should be keep
+            (loop && i === 1) ||
+            // The real last swiper should be keep
+            (loop && i === total - 1)
           ) {
             return (
               <View style={pageStyle} key={i}>
